backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
razorpay
python-dotenv
# smtplib
email-validator
